requirements.txt
anyio==3.4.0
asgiref==3.4.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0
django-cleanup==5.2.0
django-crispy-forms==1.13.0
django-heroku==0.3.1
django-phonenumber-field==6.0.0
django-storages==1.12.3
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
jmespath==0.10.0
oauthlib==3.1.1
phonenumbers==8.12.40
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
wikipedia==1.4.0
youtube-search-python==1.5.3
